ocb3

Struct Ocb3

pub struct Ocb3<Cipher, NonceSize = U12, TagSize = U16, const L_TABLE_SIZE: usize = 26>
where
    NonceSize: NonceSizes,
    TagSize: TagSizes,
{
    pub(crate) cipher: Cipher,
    pub(crate) nonce_size: PhantomData<NonceSize>,
    pub(crate) tag_size: PhantomData<TagSize>,
    pub(crate) ll_star: Array<u8, U16>,
    pub(crate) ll_dollar: Array<u8, U16>,
    pub(crate) ll: [Array<u8, U16>; L_TABLE_SIZE],
}

OCB3: generic over a block cipher implementation, nonce size, and tag size.

NonceSize: max of 15-bytes, default and recommended size of 12-bytes (96-bits). We further restrict the minimum nonce size to 6-bytes to prevent an attack described in the following paper: https://eprint.iacr.org/2023/326.pdf.

TagSize: non-zero, max of 16-bytes, default and recommended size of 16-bytes.

Compilation will fail if the size conditions are not satisfied:

// Invalid nonce size equal to 5 bytes
let cipher = Ocb3::<Aes128, U5>::new(&key);

// Invalid nonce size equal to 16 bytes
let cipher = Ocb3::<Aes128, U16>::new(&key);

// Invalid tag size equal to 0 bytes
let cipher = Ocb3::<Aes128, U12, U0>::new(&key);

// Invalid tag size equal to 20 bytes
let cipher = Ocb3::<Aes128, U12, U20>::new(&key);

Maximum plaintext and associated data lengths

The maximum lengths depend on the L_TABLE_SIZE constant which sets number of L values to be precomputed. Precomputing m values, allows processing inputs of length up to 2^m blocks (2^m * 16 bytes) without needing to calculate L values at runtime.

By default L_TABLE_SIZE is equal to 26, which allows to handle up to 1 GiB of data.

Fields
cipher: Ciphernonce_size: PhantomData<NonceSize>tag_size: PhantomData<TagSize>ll_star: Array<u8, U16>ll_dollar: Array<u8, U16>ll: [Array<u8, U16>; L_TABLE_SIZE]
Implementations

impl<Cipher, NonceSize, TagSize, const L_TABLE_SIZE: usize> Ocb3<Cipher, NonceSize, TagSize, L_TABLE_SIZE>
where Cipher: BlockSizeUser<BlockSize = U16> + BlockCipherEncrypt + BlockCipherDecrypt, NonceSize: NonceSizes, TagSize: TagSizes,

pub(crate) fn decrypt_inout_return_tag( &self, nonce: &Nonce<NonceSize>, associated_data: &[u8], buffer: InOutBuf<'_, '_, u8>, ) -> Result<Tag<Self>>

Decrypts in place and returns expected tag.

pub(crate) fn wide_encrypt<'i, 'o>( &self, nonce: &Nonce<NonceSize>, buffer: InOutBuf<'i, 'o, u8>, ) -> (InOutBuf<'i, 'o, u8>, usize, Array<u8, U16>, Array<u8, U16>)

Encrypts plaintext in groups of two.

Adapted from https://www.cs.ucdavis.edu/~rogaway/ocb/news/code/ocb.c

pub(crate) fn wide_decrypt<'i, 'o>( &self, nonce: &Nonce<NonceSize>, buffer: InOutBuf<'i, 'o, u8>, ) -> (InOutBuf<'i, 'o, u8>, usize, Array<u8, U16>, Array<u8, U16>)

Decrypts plaintext in groups of two.

Adapted from https://www.cs.ucdavis.edu/~rogaway/ocb/news/code/ocb.c

pub(crate) fn hash(&self, associated_data: &[u8]) -> Array<u8, U16>

Computes HASH function defined in https://www.rfc-editor.org/rfc/rfc7253.html#section-4.1

pub(crate) fn compute_tag( &self, associated_data: &[u8], checksum_m: &mut Array<u8, U16>, offset_m: &Array<u8, U16>, ) -> Tag<TagSize>